refactor: domain layer to specific folder and better imports
