+++ src/domain/entities/curso.ts
