+++ src/domain/entities/departamento.ts
