+++ src/domain/entities/grupoAcademico.ts
