+++ src/domain/entities/local.ts
