+++ src/domain/entities/usuario.ts
