+++ src/domain/errors/baseError.ts
