+++ src/domain/errors/index.ts
