+++ src/factories/usecase/auth/loginUsecaseFactory.ts
@@ -1,7 +1,7 @@
-import { ILoginUsecase, LoginUsecase } from 'application/usecase/user/loginUsecase'
-import { JwtAuthService } from 'infra/auth/jwtAuthService'
-import { BcryptPasswordService } from 'infra/password/bcryptPasswordService'
-import { MemoryUserRepository } from 'infra/repositories/user/memoryUserRepository'
+import { JwtAuthService } from '@infra/auth/jwtAuthService'
+import { BcryptPasswordService } from '@infra/password/bcryptPasswordService'
+import { MemoryUserRepository } from '@infra/repositories/user/memoryUserRepository'
+import { ILoginUsecase, LoginUsecase } from '@usecases/auth/loginUsecase'
 
 export class LoginUsecaseFactory {
   createUsecase(): ILoginUsecase {
