+++ src/factories/usecase/user/createUserUsecaseFactory.ts
@@ -1,7 +1,7 @@
-import { CreateUserUsecase, ICreateUserUsecase } from 'application/usecase/user/createUserUsecase'
-import { UuidIdService } from 'infra/id/uuidIdService'
-import { BcryptPasswordService } from 'infra/password/bcryptPasswordService'
-import { MemoryUserRepository } from 'infra/repositories/user/memoryUserRepository'
+import { UuidIdService } from '@infra/id/uuidIdService'
+import { BcryptPasswordService } from '@infra/password/bcryptPasswordService'
+import { MemoryUserRepository } from '@infra/repositories/user/memoryUserRepository'
+import { CreateUserUsecase, ICreateUserUsecase } from '@usecases/user/createUserUsecase'
 
 export class CreateUserUsecaseFactory {
   createUsecase(): ICreateUserUsecase {
