+++ tsconfig.json
@@ -4,18 +4,26 @@
     "module": "commonjs",
     "baseUrl": "./src",
     "paths": {
-      "@repositories/*": ["./application/repositories/*"],
-      "@entities/*": ["./application/domain/entities/*"],
+      "@domain/*": ["./domain/*"],
+      "@entities/*": ["./domain/entities/*"],
+      "@errors/*": ["./domain/errors/*"],
+      "@errors": ["./domain/errors"],
+      "@application/*": ["./application/*"],
       "@usecases/*": ["./application/usecases/*"],
-      "@infrastructure/*": ["./infrastructure/*"],
-      "@controllers/*": ["./http/controllers/*"],
-      "@middlewares/*": ["./http/middlewares/*"],
+      "@usecases": ["./application/usecases"],
+      "@repositories/*": ["./application/repositories/*"],
+      "@services/*": ["./application/services/*"],
+      "@dtos/*": ["./application/dtos/*"],
+      "@infra/*": ["./infra/*"],
       "@http/*": ["./http/*"],
       "@http": ["./http/"],
+      "@controllers/*": ["./http/controllers/*"],
+      "@middlewares/*": ["./http/middlewares/*"],
+      "@adapters/*": ["./http/adapters/*"],
+      "@routes/*": ["./http/routes/*"],
+      "@routes": ["./http/routes"],
       "@factories/*": ["./factories/*"],
-      "@utils/*": ["./utils/*"],
-      "@errors/*": ["./application/domain/errors/*"],
-      "@errors": ["./application/domain/errors"]
+      "@utils/*": ["./utils/*"]
     },
     "typeRoots": ["./src/@types"],
     "sourceMap": true,
